Move token fetching to interface layout
