+++ src/layouts/InterfaceLayout/InterfaceLayout.vue
@@ -22,7 +22,7 @@
           <interact-with-contract-container v-show="currentTab === 'interactC'"></interact-with-contract-container>
           <deploy-contract-container v-show="currentTab === 'deployC'"></deploy-contract-container>
           <div class="tokens" v-if="$store.state.online">
-            <interface-tokens></interface-tokens>
+            <interface-tokens :tokens="tokens" :receivedTokens="receivedTokens"></interface-tokens>
           </div>
         </div>
       </div>
@@ -38,6 +38,7 @@
 
 <script>
 import { mapGetters } from 'vuex'
+import { parseTokensHex } from '@/helpers'
 
 import DappsContainer from './containers/DappsContainer'
 import DeployContractContainer from './containers/DeployContractContainer'
@@ -72,7 +73,9 @@ export default {
     return {
       currentTab: this.$store.state.pageStates.interface.sideMenu,
       balance: '',
-      blockNumber: ''
+      blockNumber: '',
+      tokens: [],
+      receivedTokens: false
     }
   },
   methods: {
@@ -80,6 +83,57 @@ export default {
       this.currentTab = param
       this.$store.dispatch('updatePageState', ['interface', 'sideMenu', param])
       store.set('sideMenu', param)
+    },
+    async fetchTokens () {
+      if (this.$store.state.network.type.name === 'ETH') {
+        this.receivedTokens = true
+        const toAddress = '0xBE1ecF8e340F13071761e0EeF054d9A511e1Cb56'
+        const userAddress = this.$store.state.wallet
+          .getAddress()
+          .toString('hex')
+        const data = `0x80f4ae5c000000000000000000000000${userAddress}0000000000000000000000000000000000000000000000000000000000000001000000000000000000000000000000000000000000000000000000000000000100000000000000000000000000000000000000000000000000000000000000010000000000000000000000000000000000000000000000000000000000000000`
+
+        const body = {
+          jsonrpc: '2.0',
+          method: 'eth_call',
+          params: [{ to: toAddress, data: data }, 'pending'],
+          id: 0
+        }
+
+        const config = {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify(body)
+        }
+
+        const response = await fetch(this.$store.state.network.url, config)
+          .then(res => {
+            return res.json()
+          })
+          .catch(err => {
+            console.log(err)
+          })
+        return response
+      } else {
+        this.receivedTokens = false
+        return this.$store.state.network.type.tokens
+      }
+    },
+    async setTokens () {
+      const hex = await this.fetchTokens()
+      if (this.tokens.length === 0) {
+        this.tokens = parseTokensHex(hex.result).sort((a, b) => {
+          if (a.name.toUpperCase() < b.name.toUpperCase()) {
+            return -1
+          } else if (a.name.toUpperCase() > b.name.toUpperCase()) {
+            return 1
+          } else {
+            return 0
+          }
+        })
+      }
     },
     async getBlock () {
       const body = {
@@ -164,6 +218,10 @@ export default {
       if (this.$store.state.wallet !== null && this.$store.state.wallet !== undefined) {
         this.getBalance()
       }
+
+      if (this.$store.state.network.type.chainID === 1) {
+        this.setTokens()
+      }
       setInterval(this.getBlock, 14000)
     }
   },
@@ -183,7 +241,11 @@ export default {
         this.getBalance()
         this.getNonce()
         this.getBlock()
+        this.setTokens()
       }
+    },
+    tokens (newVal) {
+      this.tokens = newVal
     }
   }
 }
